feat: update ProdutoBuilder to use parseInt for price and enhance ProdutoController to accept token for product creation

+++ src/builders/produto.builder.ts
@@ -7,7 +7,7 @@ export class ProdutoBuilder {
   constructor() {
     this.produto = {
       nome: faker.commerce.productName(),
-      preco: parseFloat(faker.commerce.price()),
+      preco: parseInt(faker.commerce.price()),
       descricao: faker.commerce.productDescription(),
       quantidade: faker.number.int({ min: 1, max: 100 }),
     };
